Can multiple if-statements work too?
I'm sorry, I don't know how else to reach out to you about this question, and I figured this may be the best way. For challenge 3, why is it possible to do multiple if-statements? I've tried to run it, and it does not work (there is no output, just a blank), and I was wondering why. Can you explain?
Note: if I assign a value less than 10 for the variable x (i.e. I make the first if-statement true), an output shows. However, any value that makes the other if-statements true (i.e. a value between 10-25 that makes the 2nd if-statement true, or a value greater than 25 that makes the 3rd if-statement true), no output shows. Again, can you explain?