Update chap6_challenge3.py
The sample text from the book is not the same as the string in the code. The use of the title() capitalizes the first letter of each word without regard to the part of speech. Only "aldous" needs to be capitalized, so the capitalize() method is correct for what the challenge is asking for.